LPS-55772 With afterPropertiesSet gone, no need for separate classes
https://issues.liferay.com/browse/LPS-55772
